use our capture std i/o context
